add anr

+++ window_scene/session/container/src/zidl/window_event_channel_proxy.cpp
@@ -84,12 +84,12 @@ int32_t WindowEventChannelProxy::GetApplicationPid()
     MessageOption option(MessageOption::TF_SYNC);
     if (!data.WriteInterfaceToken(GetDescriptor())) {
         WLOGFE("WriteInterfaceToken failed");
-        return WSError::WS_ERROR_IPC_FAILED;
+        return static_cast<int32_t>(WSError::WS_ERROR_IPC_FAILED);
     }
     if (Remote()->SendRequest(static_cast<uint32_t>(WindowEventChannelMessage::TRANS_ID_GET_APPLICATION_PID),
         data, reply, option) != ERR_NONE) {
         WLOGFE("SendRequest failed");
-        return WSError::WS_ERROR_IPC_FAILED;
+        return static_cast<int32_t>(WSError::WS_ERROR_IPC_FAILED);
     }
     int32_t applicationPid = reply.ReadInt32();
     return applicationPid;
